Deprecated takeOrEmpty() functions in favor of orEmpty()
Updated the takeOrEmpty() functions for List, Set, Array, and Map in Collection.kt that previously returned an empty iterable if the original was null. Now they are deprecated with a suggestion to use the built-in Kotlin function orEmpty(). The built-in orEmpty() function is more idiomatic to Kotlin and improves readability. Also, minor adjustment was made on documentation comment.